requirements.txt
aiohttp
playwright
tf-playwright-stealth
python-dotenv
